Eager autoload constants for thread-safety
